+++ packages/components/src/item-group/item-group/component.tsx
@@ -7,13 +7,13 @@ import type { Ref } from 'react';
 /**
  * Internal dependencies
  */
-import type { PolymorphicComponentProps } from '../ui/context';
+import type { PolymorphicComponentProps } from '../../ui/context';
 // eslint-disable-next-line no-duplicate-imports
-import { contextConnect } from '../ui/context';
-import { useItemGroup } from './use-item-group';
-import { ItemGroupContext, useItemGroupContext } from './context';
-import { View } from '../view';
-import type { ItemGroupProps } from './types';
+import { contextConnect } from '../../ui/context';
+import { useItemGroup } from './hook';
+import { ItemGroupContext, useItemGroupContext } from '../context';
+import { View } from '../../view';
+import type { ItemGroupProps } from '../types';
 
 function ItemGroup(
 	props: PolymorphicComponentProps< ItemGroupProps, 'div' >,
