+++ packages/components/src/item-group/item-group/hook.ts
@@ -1,16 +1,16 @@
 /**
  * Internal dependencies
  */
-import { useContextSystem } from '../ui/context';
+import { useContextSystem } from '../../ui/context';
 // eslint-disable-next-line no-duplicate-imports
-import type { PolymorphicComponentProps } from '../ui/context';
+import type { PolymorphicComponentProps } from '../../ui/context';
 
 /**
  * Internal dependencies
  */
-import * as styles from './styles';
-import { useCx } from '../utils/hooks/use-cx';
-import type { ItemGroupProps } from './types';
+import * as styles from '../styles';
+import { useCx } from '../../utils/hooks/use-cx';
+import type { ItemGroupProps } from '../types';
 
 export function useItemGroup(
 	props: PolymorphicComponentProps< ItemGroupProps, 'div' >
