+++ packages/components/src/item-group/item-group/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useItemGroup } from './hook';
