+++ packages/components/src/item-group/item/component.ts
@@ -1,8 +1,8 @@
 /**
  * Internal dependencies
  */
-import { createComponent } from '../ui/utils';
-import { useItem } from './use-item';
+import { createComponent } from '../../ui/utils';
+import { useItem } from './hook';
 
 export default createComponent( {
 	useHook: useItem,
