+++ packages/components/src/item-group/item/hook.ts
@@ -7,13 +7,13 @@ import type { ElementType } from 'react';
 /**
  * Internal dependencies
  */
-import { useContextSystem } from '../ui/context';
+import { useContextSystem } from '../../ui/context';
 // eslint-disable-next-line no-duplicate-imports
-import type { PolymorphicComponentProps } from '../ui/context';
-import * as styles from './styles';
-import { useItemGroupContext } from './context';
-import { useCx } from '../utils/hooks/use-cx';
-import type { ItemProps } from './types';
+import type { PolymorphicComponentProps } from '../../ui/context';
+import * as styles from '../styles';
+import { useItemGroupContext } from '../context';
+import { useCx } from '../../utils/hooks/use-cx';
+import type { ItemProps } from '../types';
 
 export function useItem(
 	props: PolymorphicComponentProps< ItemProps, 'div' >
