+++ packages/components/src/item-group/item/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useItem } from './hook';
